refactor: Move generate.py to pygobject_stub_generator package
Split the monolithic tools/generate.py into a proper package
under src/pygobject_stub_generator with separate modules for
arguments, class info, imports, properties, type vars, and
the main stub generator. Update tools/update.py and
tools/update_all.py to invoke it via `python -m`, exclude the
package from the installed distribution, and refresh the
CONTRIBUTING docs to reflect the new invocation.
