+++ src/pygobject_stub_generator/__main__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .main import main
+
+main()
